+++ src/widgets/chart-canvas/ChartCanvas.svelte
@@ -8,7 +8,7 @@ import { fetchCharts } from '$features/charts';
 import { LayersView } from '$features/layers-panel';
 import { createVesselOverlay } from '$features/vessel-layer';
 import {
-  baseStyleUrl,
+  baseStyle,
   beforeIdFor,
   createChartOverlay,
   installSentinels,
@@ -41,7 +41,7 @@ onMount(() => {
   try {
     map = new maplibregl.Map({
       container,
-      style: baseStyleUrl(),
+      style: baseStyle(),
       center: [0, 30],
       zoom: 2,
       attributionControl: { compact: true },